added transaction history button to the app and made styling consistent between apps
