+++ app/src/components/AppHome.tsx
@@ -1,16 +1,15 @@
 'use client'
-import { cn } from '@velocitylabs-org/turtle-ui'
+import { FormContainer, HistoryLoaderSkeleton, type TransferTabOptions } from '@velocitylabs-org/turtle-ui'
 import dynamic from 'next/dynamic'
-import { useState } from 'react'
+import { Suspense, useState } from 'react'
 import TabNavigation from '@/components/TabNavigation'
 import Transfer from '@/components/Transfer'
 import useCompletedTransfers from '@/hooks/useCompletedTransfers'
 import type { TabOptions } from '@/models/transfer'
-import TransactionLoaderSkeleton from './completed/TransactionLoaderSkeleton'
-import OngoingTransfers from './ongoing/OngoingTransfers'
+import { useOngoingTransfersStore } from '@/store/ongoingTransfersStore'
 
-const TransferHistory = dynamic(() => import('@/components/completed/TransactionHistory'), {
-  loading: () => <TransactionLoaderSkeleton />,
+const TransactionHistory = dynamic(() => import('./TransactionsHistory'), {
+  loading: () => <HistoryLoaderSkeleton length={10} />,
   ssr: false,
 })
 
@@ -18,7 +17,10 @@ export default function AppHome() {
   const { completedTransfers } = useCompletedTransfers()
   const [selectedTab, setSelectedTab] = useState<TabOptions>('New')
   const hasCompletedTransfers = !!completedTransfers && completedTransfers.length > 0
-  const isDoneTabSelected = selectedTab === 'Done'
+
+  const ongoingTransfers = useOngoingTransfersStore(state => state.transfers)
+  const [newTransferInit, setNewTransferInit] = useState<TransferTabOptions>('New')
+  const isHistoryTabSelected = newTransferInit === 'History'
 
   return (
     <>
@@ -27,15 +29,27 @@ export default function AppHome() {
         setSelectedTab={setSelectedTab}
         hasCompletedTransfers={hasCompletedTransfers}
       />
-      <div className={cn('z-15 relative max-w-[90vw]', isDoneTabSelected && 'hidden')}>
-        <Transfer />
-        <OngoingTransfers
-          selectedTab={selectedTab}
-          setSelectedTab={setSelectedTab}
-          hasCompletedTransfers={hasCompletedTransfers}
-        />
-      </div>
-      {isDoneTabSelected && <TransferHistory transfers={completedTransfers!} />}
+      <FormContainer
+        ongoingTransfers={ongoingTransfers}
+        completedTransfers={completedTransfers}
+        isHistoryTabSelected={isHistoryTabSelected}
+        newTransferInit={newTransferInit}
+        setNewTransferInit={setNewTransferInit}
+      >
+        {!isHistoryTabSelected ? (
+          <Transfer />
+        ) : (
+          <Suspense
+            fallback={
+              <HistoryLoaderSkeleton
+                length={ongoingTransfers.length + (completedTransfers ? completedTransfers.length : 0)}
+              />
+            }
+          >
+            <TransactionHistory transfers={completedTransfers} />
+          </Suspense>
+        )}
+      </FormContainer>
     </>
   )
 }
