--- app/src/components/ui/skeleton.tsx
@@ -1,7 +1,0 @@
-import { cn } from '@velocitylabs-org/turtle-ui'
-
-function Skeleton({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
-  return <div className={cn('animate-pulse rounded-md bg-turtle-level2', className)} {...props} />
-}
-
-export { Skeleton }
