+++ packages/ui/src/components/HistoryLoaderSkeleton.tsx
@@ -1,8 +1,12 @@
 import { colors } from '@velocitylabs-org/turtle-tailwind-config'
-import { ArrowRight } from '@/assets/svg/ArrowRight'
-import { Skeleton } from '@/components/ui/skeleton'
+import { cn } from '@velocitylabs-org/turtle-ui'
+import { ArrowRight } from 'lucide-react'
 
-const SkeletonAddresses = () => {
+function Skeleton({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
+  return <div className={cn('animate-pulse rounded-md bg-turtle-level2', className)} {...props} />
+}
+
+function SkeletonAddresses() {
   return (
     <div className="flex items-center justify-start space-x-4">
       {/* Transaction skeleton sender */}
@@ -10,7 +14,7 @@ const SkeletonAddresses = () => {
         <Skeleton className="h-4 w-4 rounded-full border-turtle-level3" />
         <Skeleton className="h-5 w-20" />
       </div>
-      <ArrowRight className="h-3 w-3" {...{ fill: colors['turtle-level2'] }} />
+      <ArrowRight className="h-3 w-3" fill={colors['turtle-level2']} />
       {/* Transaction skeleton recipient */}
       <div className="flex items-center gap-x-2">
         <Skeleton className="h-4 w-4 rounded-full border-turtle-level3" />
@@ -20,17 +24,17 @@ const SkeletonAddresses = () => {
   )
 }
 
-const SkeletonTokens = () => {
+function SkeletonTokens() {
   return (
     <div className="items-centern flex space-x-2 rounded-2xl border px-1 py-0.5">
       <Skeleton className="h-4 w-4 rounded-full" />
-      <ArrowRight className="w-[0.45rem h-[0.45rem]" {...{ fill: colors['turtle-level2'] }} />
+      <ArrowRight className="w-[0.45rem h-[0.45rem]" fill={colors['turtle-level2']} />
       <Skeleton className="h-4 w-4 rounded-full" />
     </div>
   )
 }
 
-const SkeletonCard = () => {
+function SkeletonCard() {
   return (
     <div className="flex items-center rounded-2xl border border-turtle-level3 p-4 sm:gap-4">
       <div className="w-full space-y-2">
@@ -54,14 +58,19 @@ const SkeletonCard = () => {
   )
 }
 
-const HistoryLoaderSkeleton = ({ length }: { length: number }) => {
+export const HistoryLoaderSkeleton = ({ length }: { length: number }) => {
   return (
-    <div className="z-20 flex max-h-[70vh] max-w-[90vw] flex-col gap-1 overflow-auto rounded-2xl border bg-turtle-background p-5 px-[1.5rem] py-[2rem] sm:w-[31.5rem] sm:p-[2.5rem]">
+    <div className="z-20 border border-turtle-foreground mb-12 flex flex-col gap-4 rounded-3xl bg-white p-4 px-[1.5rem] py-[2rem] sm:w-[31.5rem] sm:p-[2.5rem]">
       {Array.from({ length }, (_, i) => i + 1).map(idx => (
-        <SkeletonCard key={idx} />
+        <div key={idx} className="z-30">
+          <div className="w-full space-y-4">
+            <Skeleton className="h-5 w-1/3" />
+            {Array.from({ length: 2 }, (_, i) => i + 1).map(idx => (
+              <SkeletonCard key={idx} />
+            ))}
+          </div>
+        </div>
       ))}
     </div>
   )
 }
-
-export default HistoryLoaderSkeleton
