+++ packages/ui/src/components/index.ts
@@ -1,9 +1,10 @@
 export * from './Button'
 export * from './Checkbox'
+export * from './FormContainer'
+export * from './HistoryLoaderSkeleton'
 export * from './Icon'
 export * from './LoadingIcon'
 export * from './Switch'
 export * from './TokenLogo'
 export * from './Tooltip'
-export * from './TxHistoryToggle'
 export * from './typography'
