+++ widget/src/components/Widget.tsx
@@ -1,4 +1,4 @@
-import { type TransferTabOptions, TxHistoryToggle } from '@velocitylabs-org/turtle-ui'
+import { FormContainer, HistoryLoaderSkeleton, type TransferTabOptions } from '@velocitylabs-org/turtle-ui'
 import { lazy, Suspense, useMemo, useState } from 'react'
 import NotificationSystem from '@/components/NotificationSystem'
 import SubstrateWalletModal from '@/components/SubstrateWalletModal'
@@ -8,7 +8,6 @@ import { Providers } from '@/providers'
 import { ConfigProvider, type ConfigRegistryType } from '@/providers/ConfigProviders'
 import { useOngoingTransfersStore } from '@/stores/ongoingTransfersStore'
 import { generateWidgetTheme, type WidgetTheme } from '@/utils/theme'
-import HistoryLoaderSkeleton from './history/HistoryLoaderSkeleton'
 
 const Widget = ({ theme, registry }: { theme?: WidgetTheme; registry?: ConfigRegistryType }) => {
   useMemo(() => generateWidgetTheme(theme), [theme])
@@ -25,32 +24,31 @@ const Widget = ({ theme, registry }: { theme?: WidgetTheme; registry?: ConfigReg
     <div className="turtle-wrapper">
       <Providers>
         <ConfigProvider registry={registry ?? { chains: [], tokens: [] }}>
-          <div className="m-4 flex flex-col items-center justify-center p-6">
-            <div className="relative">
-              <TxHistoryToggle
-                ongoingTransfers={ongoingTransfers}
-                completedTransfers={completedTransfers}
-                isHistoryTabSelected={isHistoryTabSelected}
-                newTransferInit={newTransferInit}
-                setNewTransferInit={setNewTransferInit}
-              />
-              {!isHistoryTabSelected ? (
-                <TransferForm />
-              ) : (
-                <Suspense
-                  fallback={
+          <FormContainer
+            ongoingTransfers={ongoingTransfers}
+            completedTransfers={completedTransfers}
+            isHistoryTabSelected={isHistoryTabSelected}
+            newTransferInit={newTransferInit}
+            setNewTransferInit={setNewTransferInit}
+          >
+            {!isHistoryTabSelected ? (
+              <TransferForm />
+            ) : (
+              <Suspense
+                fallback={
+                  <div className="z-20 flex flex-col gap-1 overflow-auto rounded-3xl border border-turtle-foreground bg-turtle-background">
                     <HistoryLoaderSkeleton
                       length={ongoingTransfers.length + (completedTransfers ? completedTransfers.length : 0)}
                     />
-                  }
-                >
-                  <TransfersHistory ongoingTransfers={ongoingTransfers} completedTransfers={completedTransfers ?? []} />
-                </Suspense>
-              )}
-            </div>
+                  </div>
+                }
+              >
+                <TransfersHistory ongoingTransfers={ongoingTransfers} completedTransfers={completedTransfers ?? []} />
+              </Suspense>
+            )}
             <NotificationSystem />
             <SubstrateWalletModal />
-          </div>
+          </FormContainer>
         </ConfigProvider>
       </Providers>
     </div>
