+++ widget/src/components/history/TransfersHistory.tsx
@@ -21,7 +21,7 @@ const TransfersHistory = ({
   return (
     <div
       id="ongoing-txs"
-      className="z-20 flex max-h-[70vh] max-w-[90vw] flex-col gap-1 overflow-auto rounded-2xl border bg-turtle-background p-5 px-[1.5rem] py-[2rem] sm:w-[31.5rem] sm:p-[2.5rem]"
+      className="z-20 flex max-h-[70vh] max-w-[90vw] flex-col gap-1 overflow-auto rounded-3xl border border-turtle-foreground bg-turtle-background p-5 px-[1.5rem] py-[2rem] sm:w-[31.5rem] sm:p-[2.5rem]"
     >
       {!ongoingTransfers.length && !completedTransfers.length ? (
         <div className="text-center text-sm text-turtle-level5">No transfers yet</div>
