+++ widget/src/components/history/ongoing-transfers/Card.tsx
@@ -11,7 +11,6 @@ const OngoingTransfer: FC<{
   transfer: StoredTransfer
 }> = ({ transfer }) => {
   return (
-    //prettier-ignore
     <div className="hover:bg-turtle-secondary-light-50 mb-2 space-y-2 rounded-[16px] border border-turtle-level3 p-3 hover:cursor-pointer">
       <div className="flex items-center justify-between">
         <div className="flex items-center">
